+++ lib/rpcs/results.rs
@@ -25,6 +25,7 @@ pub use super::v2_0_0::list_rpcs::{
     SchemaParam,
 };
 pub use super::v2_0_0::put_deploy::PutDeployResult;
+pub use super::v2_0_0::put_transaction::PutTransactionResult;
 pub use super::v2_0_0::query_balance::QueryBalanceResult;
 pub use super::v2_0_0::query_global_state::QueryGlobalStateResult;
 pub use super::v2_0_0::speculative_exec::SpeculativeExecResult;