+++ src/transaction.rs
@@ -1,6 +1,8 @@
 mod creation_common;
 mod make;
+mod put;
 mod sign;
 
 pub use make::MakeTransaction;
+pub use put::PutTransaction;
 pub use sign::SignTransaction;